+++ app/src/main/java/fr/free/nrw/commons/contributions/ContributionsListFragment.java
@@ -1,11 +1,17 @@
 package fr.free.nrw.commons.contributions;
 
+import android.Manifest;
 import android.app.Activity;
 import android.content.ActivityNotFoundException;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
+import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -24,6 +30,7 @@ import fr.free.nrw.commons.AboutActivity;
 import fr.free.nrw.commons.CommonsApplication;
 import fr.free.nrw.commons.R;
 import fr.free.nrw.commons.SettingsActivity;
+import fr.free.nrw.commons.upload.UploadService;
 
 public class ContributionsListFragment extends Fragment {
 
@@ -83,9 +90,15 @@ public class ContributionsListFragment extends Fragment {
 
     @Override
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        //FIXME: must get the file data for Google Photos when receive the intent answer, in the onActivityResult method
         super.onActivityResult(requestCode, resultCode, data);
-        if(resultCode == Activity.RESULT_OK) {
-            controller.handleImagePicked(requestCode, data);
+
+        if (data != null) {
+            Log.d("Contributions", "OnActivityResult() parameters: Result code: " + resultCode + " Data: " + data.toString());
+            Uri imageData = data.getData();
+            controller.handleImagePicked(requestCode, imageData);
+        } else {
+            Log.e("Contributions", "OnActivityResult() parameters: Result code: " + resultCode + " Data: null");
         }
     }
 
@@ -94,8 +107,20 @@ public class ContributionsListFragment extends Fragment {
     public boolean onOptionsItemSelected(MenuItem item) {
         switch(item.getItemId()) {
             case R.id.menu_from_gallery:
-                controller.startGalleryPick();
-                return true;
+                //Gallery crashes before reach ShareActivity screen so must implement permissions check here
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                    if (ContextCompat.checkSelfPermission(this.getActivity(), Manifest.permission.READ_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
+                        ActivityCompat.requestPermissions(this.getActivity(), new String[]{Manifest.permission.READ_EXTERNAL_STORAGE}, 1);
+                        return true;
+                    } else {
+                        controller.startGalleryPick();
+                        return true;
+                    }
+                }
+                else {
+                    controller.startGalleryPick();
+                    return true;
+                }
             case R.id.menu_from_camera:
                 controller.startCameraCapture();
                 return true;
@@ -126,6 +151,23 @@ public class ContributionsListFragment extends Fragment {
                 return true;
             default:
                 return super.onOptionsItemSelected(item);
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode,
+                                           String permissions[], int[] grantResults) {
+        switch (requestCode) {
+            // 1 = Storage allowed when gallery selected
+            case 1: {
+                if (grantResults.length > 0
+                        && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                    controller.startGalleryPick();
+                } else {
+                    return;
+                }
+                return;
+            }
         }
     }