+++ app/src/main/java/fr/free/nrw/commons/upload/GPSExtractor.java
@@ -57,15 +57,25 @@ public class GPSExtractor {
         provider = locationManager.getBestProvider(criteria, true);
         myLocationListener = new MyLocationListener();
 
-        locationManager.requestLocationUpdates(provider, 400, 1, myLocationListener);
-        Location location = locationManager.getLastKnownLocation(provider);
-        if (location != null) {
-            myLocationListener.onLocationChanged(location);
+        try {
+            locationManager.requestLocationUpdates(provider, 400, 1, myLocationListener);
+            Location location = locationManager.getLastKnownLocation(provider);
+            if (location != null) {
+                myLocationListener.onLocationChanged(location);
+            }
+        } catch (IllegalArgumentException e) {
+            Log.e(TAG, "Illegal argument exception", e);
+        } catch (SecurityException e) {
+            Log.e(TAG, "Security exception", e);
         }
     }
 
     protected void unregisterLocationManager() {
-        locationManager.removeUpdates(myLocationListener);
+        try {
+            locationManager.removeUpdates(myLocationListener);
+        } catch (SecurityException e) {
+            Log.e(TAG, "Security exception", e);
+        }
     }
 
     /**